Remove guides from acceptance tests
